Drop an unnecessary inode conversion in the artifact-notify test

+++ internal/engine/artifact_notify_test.go
@@ -211,5 +211,5 @@ func inodeOf(t *testing.T, path string) uint64 {
 	if !ok {
 		t.Skip("no inode information on this platform")
 	}
-	return uint64(sys.Ino)
+	return sys.Ino
 }
